dnc_data_pipeline/requirements.txt
# =============================================================================
# DNC Data Pipeline - Python Dependencies
# =============================================================================

# Core Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Apache Airflow Core
apache-airflow>=2.7.0,<3.0.0

# Database Drivers
psycopg2-binary>=2.9.0,<3.0.0

# Data Quality and Validation
great-expectations>=0.17.0,<1.0.0
pydantic>=2.0.0,<3.0.0

# File Processing
openpyxl>=3.1.0,<4.0.0
xlrd>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Logging and Monitoring
structlog>=23.0.0,<24.0.0
colorama>=0.4.0,<1.0.0

# Testing and Development
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.12.0,<6.0.0

# Documentation
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.0.0,<10.0.0
